refactor: move functions to data fetcher.ts
